+++ kwcoco/cli/coco_split.py
@@ -19,11 +19,17 @@ class CocoSplitCLI(object):
         """
         default = {
             'src': scfg.Value(None, help='input dataset to split', position=1),
-            'dst1': scfg.Value('split1.mscoco.json', help='output path1'),
-            'dst2': scfg.Value('split2.mscoco.json', help='output path2'),
+            'dst1': scfg.Value('split1.kwcoco.json', help='output path of the larger split'),
+            'dst2': scfg.Value('split2.kwcoco.json', help='output path of the smaller split'),
             'factor': scfg.Value(3, help='ratio of items put in dset1 vs dset2'),
             'rng': scfg.Value(None, help='random seed'),
             'balance_categories': scfg.Value(True, help='if True tries to balance annotation categories across splits'),
+            'write_all': scfg.Value(False, isflag=True, help=ub.paragraph(
+                '''
+                if True then all ``factor`` splits are written to disk. In this
+                case the dst1 and dst2 datasets must contain a {} format string
+                specifier so each of the output filesnames can be indexed.
+                ''')),
             'splitter': scfg.Value(
                 'auto', help=ub.paragraph(
                     '''
@@ -36,7 +42,9 @@ class CocoSplitCLI(object):
         }
         epilog = """
         Example Usage:
-            kwcoco split --src special:shapes8 --dst1=learn.mscoco.json --dst2=test.mscoco.json --factor=3 --rng=42
+            kwcoco split --src special:shapes8 --dst1=learn.kwcoco.json --dst2=test.kwcoco.json --factor=3 --rng=42
+
+            kwcoco split --src special:shapes8 --dst1="train_{03:d}.kwcoco.json" --dst2="vali_{0:3d}.kwcoco.json" --factor=3 --rng=42
         """
 
     @classmethod
@@ -55,6 +63,7 @@ class CocoSplitCLI(object):
         """
         import kwcoco
         import kwarray
+        import warnings
         from kwcoco.util import util_sklearn
 
         config = cls.CLIConfig(kw, cmdline=cmdline)
@@ -62,6 +71,21 @@ class CocoSplitCLI(object):
 
         if config['src'] is None:
             raise Exception('must specify source: {}'.format(config['src']))
+
+        if config['write_all']:
+            if not set(config['dst1']).issuperset(set('{}')):
+                raise Exception(
+                    'when write_all is True dst1 and dst2 must contain a {} format string placeholder')
+
+            if not set(config['dst2']).issuperset(set('{}')):
+                raise Exception(
+                    'when write_all is True dst1 and dst2 must contain a {} format string placeholder')
+        else:
+            if set(config['dst1']).issuperset(set('{}')):
+                warnings.warn('write_all is False, but dst1 has a format string placeholder')
+
+            if set(config['dst2']).issuperset(set('{}')):
+                warnings.warn('write_all is False, but dst2 has a format string placeholder')
 
         print('reading fpath = {!r}'.format(config['src']))
         dset = kwcoco.CocoDataset.coerce(config['src'])
@@ -102,7 +126,8 @@ class CocoSplitCLI(object):
         rng = kwarray.ensure_rng(config['rng'])
 
         shuffle = rng is not None
-        self = util_sklearn.StratifiedGroupKFold(n_splits=config['factor'],
+        factor = config['factor']
+        self = util_sklearn.StratifiedGroupKFold(n_splits=factor,
                                                  random_state=rng,
                                                  shuffle=shuffle)
 
@@ -110,25 +135,34 @@ class CocoSplitCLI(object):
             split_idxs = list(self.split(X=final_group_gids, y=final_group_cids, groups=final_group_ids))
         else:
             split_idxs = list(self.split(X=final_group_gids, y=final_group_gids, groups=final_group_ids))
-        idxs1, idxs2 = split_idxs[0]
-
-        gids1 = sorted(ub.unique(ub.take(final_group_gids, idxs1)))
-        gids2 = sorted(ub.unique(ub.take(final_group_gids, idxs2)))
-
-        dset1 = dset.subset(gids1)
-        dset2 = dset.subset(gids2)
 
         dumpkw = {
             'newlines': True,
             'compress': config['compress'],
         }
-        dset1.fpath = config['dst1']
-        print('Writing dset1 = {!r}'.format(dset1.fpath))
-        dset1.dump(**dumpkw)
+        for split_num, (idxs1, idxs2) in enumerate(split_idxs):
+            idxs1, idxs2 = split_idxs[0]
+            gids1 = sorted(ub.unique(ub.take(final_group_gids, idxs1)))
+            gids2 = sorted(ub.unique(ub.take(final_group_gids, idxs2)))
 
-        dset2.fpath = config['dst2']
-        print('Writing dset2 = {!r}'.format(dset2.fpath))
-        dset2.dump(**dumpkw)
+            dset1 = dset.subset(gids1)
+            dset2 = dset.subset(gids2)
+
+            if config['write_all']:
+                dset1.fpath = config['dst1'].format(split_num)
+                dset2.fpath = config['dst2'].format(split_num)
+                print(f'Writing dset1({split_num} / {factor}) = {dset1.fpath!r}')
+                dset1.dump(**dumpkw)
+                print(f'Writing dset2({split_num} / {factor}) = {dset2.fpath!r}')
+                dset2.dump(**dumpkw)
+            else:
+                dset1.fpath = config['dst1']
+                dset2.fpath = config['dst2']
+                print(f'Writing dset1 = {dset1.fpath!r}')
+                dset1.dump(**dumpkw)
+                print(f'Writing dset2 = {dset2.fpath!r}')
+                dset2.dump(**dumpkw)
+                break
 
 _CLI = CocoSplitCLI